flask/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
Werkzeug==1.0.1
celery==4.4.7
redis==3.5.3